Make oq export work in multi-user context [skip hazardlib][demos]
