+++ openquake/commands/export.py
@@ -18,7 +18,8 @@
 import os
 import logging
 
-from openquake.baselib import general, performance, sap, datastore
+from openquake.baselib import general, performance, sap
+from openquake.engine import engine
 from openquake.calculators.export import export as export_
 
 
@@ -29,7 +30,7 @@ def export(datastore_key, calc_id=-1, exports='csv', export_dir='.'):
     Export an output from the datastore.
     """
     logging.basicConfig(level=logging.INFO)
-    dstore = datastore.read(calc_id)
+    dstore = engine.read(calc_id)
     parent_id = dstore['oqparam'].hazard_calculation_id
     if parent_id:
         dstore.parent = datastore.read(parent_id)
