+++ openquake/commands/show.py
@@ -26,20 +26,10 @@ from openquake.baselib import datastore
 from openquake.commonlib.writers import write_csv
 from openquake.commonlib.util import rmsep
 from openquake.commonlib import logs
+from openquake.engine import engine
 from openquake.calculators import getters
 from openquake.calculators.views import view
 from openquake.calculators.extract import extract
-
-if config.dbserver.multi_user:
-    def read(calc_id):
-        job = logs.dbcmd('get_job', calc_id)
-        if job:
-            return datastore.read(job.ds_calc_dir + '.hdf5')
-        # calc_id can be present in the datastore and not in the database:
-        # this happens if the calculation was run with `oq run`
-        return datastore.read(calc_id)
-else:  # get the datastore of the current user
-    read = datastore.read
 
 
 def get_hcurves_and_means(dstore):
@@ -67,7 +57,7 @@ def show(what='contents', calc_id=-1, extra=()):
         rows = []
         for calc_id in datastore.get_calc_ids(datadir):
             try:
-                ds = read(calc_id)
+                ds = datastore.read(calc_id)
                 oq = ds['oqparam']
                 cmode, descr = oq.calculation_mode, oq.description
             except Exception:
@@ -82,7 +72,7 @@ def show(what='contents', calc_id=-1, extra=()):
             print('#%d %s: %s' % row)
         return
 
-    ds = read(calc_id)
+    ds = engine.read(calc_id)
 
     # this part is experimental
     if what == 'rlzs' and 'poes' in ds:
